removal of calculation function with arbitrary precision

+++ database/migrations/2020_07_19_130746_create_table__transactions.php
@@ -16,7 +16,7 @@ class CreateTableTransactions extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->decimal('value');
+            $table->float('value', 16, 2);
             $table->string('status', 30);
 
             $table->bigInteger('payer_id')->unsigned();
